+++ CHANGELOG.md
@@ -6,7 +6,8 @@ _(add items here for easier creation of next log entry)_
 
 - Test the typeahead with end to end tests.
 - Don't display hints on browsers that don't support pointer-events.
-- [Breaking] Rename `dist/styled.min.css` to `dist/accessible-typeahead.min.css`
+- [Breaking] Rename `dist/styled.min.css` to `dist/accessible-typeahead.min.css`.
+- [Breaking] Rename library main export from `AccessibleTypeahead` to `accessibleTypeahead`
 
 ## 0.4.2 - 2017-05-03